requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
slowapi>=0.1.9
pydantic>=2.0
cryptography>=44.0.0
psycopg2-binary>=2.9.0
